+++ core/components/contextkeeper/lexicon/de/setting.inc.php
@@ -9,5 +9,7 @@ $_lang['setting_contextkeeper.check_empty'] = 'Leer prüfen';
 $_lang['setting_contextkeeper.check_empty_desc'] = 'Testen Sie die Schreibbeschränkung auch dann, wenn die System-/Benutzergruppen-/Benutzereinstellung `writable_contexts` leer ist.';
 $_lang['setting_contextkeeper.debug'] = 'Debug';
 $_lang['setting_contextkeeper.debug_desc'] = 'Log debug information in the MODX error log.';
+$_lang['setting_contextkeeper.disable_duplicate_button'] = 'Duplizieren Button deaktivieren';
+$_lang['setting_contextkeeper.disable_duplicate_button_desc'] = 'Duplizieren Button in Ressourcen von nicht beschreibbaren Kontexten deaktivieren.';
 $_lang['setting_contextkeeper.writable_contexts'] = 'Beschreibbare Kontexte';
 $_lang['setting_contextkeeper.writable_contexts_desc'] = 'Kommagetrennte Liste der Schlüssel beschreibbarer Kontexte.';