+++ core/components/contextkeeper/lexicon/en/setting.inc.php
@@ -9,5 +9,7 @@ $_lang['setting_contextkeeper.check_empty'] = 'Check Empty';
 $_lang['setting_contextkeeper.check_empty_desc'] = 'Test the write restriction even if the the system/usergroup/user setting `writable_contexts` is empty.';
 $_lang['setting_contextkeeper.debug'] = 'Debug';
 $_lang['setting_contextkeeper.debug_desc'] = 'Log debug information in the MODX error log.';
+$_lang['setting_contextkeeper.disable_duplicate_button'] = 'Disable Duplicate Button';
+$_lang['setting_contextkeeper.disable_duplicate_button_desc'] = 'Disable the duplicate button in resources of not writable contexts.';
 $_lang['setting_contextkeeper.writable_contexts'] = 'Writable Contexts';
 $_lang['setting_contextkeeper.writable_contexts_desc'] = 'Comma separated list of keys of writable contexts.';