+++ core/components/contextkeeper/docs/changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2022-05-04
+
+### Added
+
+- Disable the save button and optionally the duplicate button on the resource edit page
+- Additional OnDocPublished/OnDocUnpublished events for handling publishing/unpublishing in the resource tree
+
 ## [1.0.0] - 2022-04-29
 
 ### Added